Issue #50 fix: Code restructure in data-products module
