+++ data-products/src/main/resources/masterdata-indexer.conf
@@ -3,8 +3,10 @@ env=local
 
 redis.host="localhost"
 redis.port="6379"
-cloudStorage.container="obsrv-data"
-cloudStorage.provider="aws"
+redis_scan_count=1000
+redis_maxPipelineSize=1000
+cloudStorage.container="/home/sankethika/obsrv-data"
+cloudStorage.provider="local"
 cloudStorage.accountName="obsrv" # Is required when azure is provider. Will only be used when azure is the provider
 druid.indexer.url="http://localhost:8888/druid/indexer/v1/task"
 druid.datasource.delete.url="http://localhost:8888/druid/coordinator/v1/datasources/"
