--- data-products/src/main/scala/org/sunbird/obsrv/dataproducts/exception/InvalidInputException.scala
@@ -1,6 +1,0 @@
-package org.sunbird.obsrv.dataproducts.exception
-
-case class InvalidInputException(msg: String, ex: Exception = null) extends Exception(msg, ex) {
-  //print the error message
-  println(msg)
-}
