+++ data-products/src/main/scala/org/sunbird/obsrv/dataproducts/helper/BaseMetricsHelper.scala
@@ -2,12 +2,12 @@ package org.sunbird.obsrv.dataproducts.helper
 
 import com.typesafe.config.Config
 import org.sunbird.obsrv.core.util.JSONUtil
-import org.sunbird.obsrv.dataproducts.model.{IJobMetric, IMetricsHelper}
+import org.sunbird.obsrv.dataproducts.model.{Actor, Context, Edata, IJobMetric, JobMetric, MetricObject, Pdata}
 import org.sunbird.obsrv.helpers.KafkaMessageProducer
 
-abstract class BaseMetricHelper(config: Config) extends IMetricsHelper {
+case class BaseMetricHelper(config: Config) {
 
-  override val metrics: Map[String, String] = Map(
+  val metrics: Map[String, String] = Map(
     "total_dataset_count" -> "total_dataset_count",
     "success_dataset_count" -> "success_dataset_count",
     "failure_dataset_count" -> "failure_dataset_count",
@@ -15,11 +15,28 @@ abstract class BaseMetricHelper(config: Config) extends IMetricsHelper {
     "total_time_taken" -> "total_time_taken"
   )
 
+  def getMetricName(name: String) = {
+    metrics.get(name).getOrElse("")
+  }
+
   val metricsProducer = KafkaMessageProducer(config)
 
   def sync(metric: IJobMetric): Unit = {
     val metricStr = JSONUtil.serialize(metric)
     metricsProducer.sendMessage(message = metricStr)
+  }
+
+  private def getObject(datasetId: String) = {
+    MetricObject(id = datasetId, `type` = "Dataset", ver = "1.0.0")
+  }
+
+  def generate(datasetId: String, edata: Edata) = {
+    val `object` = getObject(datasetId)
+    val actor = Actor(id = "MasterDataProcessorIndexerJob", `type` = "SYSTEM")
+    val pdata = Pdata(id = "DataProducts", pid = "MasterDataProcessorIndexerJob", ver = "1.0.0")
+    val context = Context(env = config.getString("env"), pdata = pdata)
+    val metric = JobMetric(actor = actor, context = context, `object` = `object`, edata = edata)
+    this.sync(metric)
   }
 }
