+++ data-products/src/main/scala/org/sunbird/obsrv/dataproducts/helper/KafkaMessageProducer.scala
@@ -2,6 +2,7 @@ package org.sunbird.obsrv.helpers
 
 import com.typesafe.config.Config
 import org.apache.kafka.clients.producer.{KafkaProducer, ProducerRecord}
+import org.slf4j.LoggerFactory
 
 import java.util.Properties
 
@@ -10,6 +11,7 @@ case class KafkaMessageProducer(config: Config) {
   private val kafkaProperties = new Properties();
   private val defaultTopicName = config.getString("metrics.topicName")
   private val defaultKey = null
+  private[this] val logger = LoggerFactory.getLogger(KafkaMessageProducer.getClass)
 
   kafkaProperties.put("bootstrap.servers", config.getString("kafka.bootstrap.servers"))
   kafkaProperties.put("key.serializer", "org.apache.kafka.common.serialization.StringSerializer")
@@ -23,7 +25,8 @@ case class KafkaMessageProducer(config: Config) {
       producer.send(record)
     } catch {
       case e: Exception =>
-        println("Error sending message to Kafka", e.getMessage)
+        logger.error("Exception while sending message to Kafka")
+        logger.error("Exception - "+ e.getMessage)
     }
   }
 }
