--- data-products/src/main/scala/org/sunbird/obsrv/dataproducts/model/JobConfig.scala
@@ -1,10 +1,0 @@
-package org.sunbird.obsrv.dataproducts.model
-
-import com.typesafe.config.Config
-
-trait JobConfig {
-  val name: String;
-  val config: Config;
-  val args: Array[String]
-  val postgresConfig: Map[String, AnyRef]
-}
